# internal/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: Openspace OS Core REST API
  version: "1.0.0"
  description: |
    Openspace OS Core 是面向航天运营的开放平台核心服务，提供节点管理、关系图、事件总线、
    指令流水线、遥测流水线、插件管理等能力。

    本文档描述 Openspace OS Core 对外暴露的 REST API（前缀 `/api/v1`）。
    除 `/api/v1/auth/login`、`/api/v1/health`、`/api/v1/docs`、`/api/v1/openapi.yaml`、
    `/healthz`、`/metrics` 外，其余端点在启用认证时需要携带 `Authorization: Bearer <token>` 头部。
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Openspace OS
    url: https://github.com/jeffreytanhao-eng/openspace-os-core
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: http://localhost:9090
    description: 本地开发环境（备用端口）
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新 token
      description: 使用当前有效的 token 换取新的 token。需携带 Authorization 头。
      operationId: refreshToken
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: token 无效或缺失
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 认证服务未启用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: 列出用户
      description: 返回所有用户列表。需要 user.manage 权限（admin）。
      operationId: listUsers
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 认证服务未启用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Users]
      summary: 创建用户
      description: 创建新用户。需要 user.manage 权限（admin）。
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              username: operator1
              password: secret
              roles: [operator]
              communityId: comm-1
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userId}:
    delete:
      tags: [Users]
      summary: 删除用户
      description: 删除指定用户。需要 user.manage 权限（admin）。
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: 删除成功
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/token:
    post:
      tags: [Auth]
      summary: API Key 换取 JWT
      description: |
        机器接入方（Client）用其 API Key 换取短期 JWT。
        请求体可直接携带 `apiKey`（aos_ 前缀），或通过 `Authorization: Bearer <apiKey>` 头传递。
        此后使用返回的 JWT 作为 `Authorization: Bearer <token>` 调用其余端点。此端点无需权限。
      security: []
      operationId: exchangeToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenExchangeRequest'
      responses:
        '200':
          description: 换取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: API Key 无效或被吊销
  /api/v1/clients:
    get:
      tags: [Auth]
      summary: 列出所有机器接入方
      description: 需 admin 权限（user.manage）。
      operationId: listClients
      responses:
        '200':
          description: 机器接入方列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      tags: [Auth]
      summary: 创建机器接入方并生成 API Key
      description: 需 admin 权限（user.manage）。仅创建响应返回明文 API Key，此后不可再获取。
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: 创建成功，含明文 API Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: 参数校验失败
  /api/v1/clients/{clientId}:
    get:
      tags: [Auth]
      summary: 查询单个机器接入方
      description: 需 admin 权限（user.manage）。
      operationId: getClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 机器接入方详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: 未找到
    delete:
      tags: [Auth]
      summary: 吊销机器接入方
      description: 吊销后其 API Key 与 JWT 均失效。需 admin 权限（user.manage）。
      operationId: revokeClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 吊销成功
        '404':
          description: 未找到

  # ===== 遥测上报（T5） =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "使用 JWT Bearer token 认证。先调用 /api/v1/auth/login 获取 token，再以 `Authorization: Bearer <token>` 形式携带。"
  parameters:
...
    UserId:
      name: userId
      in: path
      description: 用户 ID
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      description: 统一错误响应
      properties:
        error:
          type: string
          description: 错误码（HTTP 状态文本）
          example: not_found
        message:
          type: string
          description: 错误详情
          example: "节点不存在: nodeId=sat-1"
      required: [error, message]
...
    LoginResponse:
      type: object
      description: 登录成功响应
      properties:
        token:
          type: string
          description: JWT token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: 有效期（秒）
          example: 86400
      required: [token, expiresIn]
    User:
      type: object
      description: 用户信息
      properties:
        userId:
          type: string
          description: 用户 ID
        username:
          type: string
          description: 用户名
          example: admin
        roles:
          type: array
          description: 角色列表
          items:
            type: string
          example: [admin]
        communityId:
          type: string
          description: 所属社区 ID
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    CreateUserRequest:
      type: object
      description: 创建用户请求体
      properties:
        username:
          type: string
          example: operator1
        password:
          type: string
          format: password
          example: secret
        roles:
          type: array
          items:
            type: string
          example: [operator]
        communityId:
          type: string
          example: comm-1
      required: [username, password, roles]
...
    CreateClientRequest:
      type: object
      description: 创建机器接入方请求体
      properties:
        name:
          type: string
          example: demo-satellite-watcher
        moduleName:
          type: string
          example: watcher
        communityId:
          type: string
          example: community-001
        roles:
          type: array
          items:
            type: string
          example: [operator]
      required: [name]
    Client:
      type: object
      description: 机器接入方（Client）信息
      properties:
        clientId:
          type: string
          description: 客户端 ID
        name:
          type: string
          description: 客户端名称
        moduleName:
          type: string
          description: 所属模块名
        communityId:
          type: string
          description: 所属社区/租户 ID
        roles:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, revoked]
          description: 状态（active 仅创建后即刻或短暂有效，吊销后为 revoked）
        apiKey:
          type: string
          description: 明文 API Key（格式 aos_...，仅创建响应返回，此后不可再获取）
          example: aos_xxxxxxxxxxxxxxxx
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TokenExchangeRequest:
      type: object
      description: API Key 换取 JWT 请求体。api_key 与 Authorization 头二选一。
      properties:
        apiKey:
          type: string
          description: 机器接入方 API Key（aos_ 前缀）
          example: aos_xxxxxxxxxxxxxxxx
      required: [apiKey]

    # ===== 遥测上报（T5）相关 Schema =====
...